Authentication is created successfully

+++ app/Http/Controllers/ThreadController.php
@@ -84,11 +84,15 @@ class ThreadController extends Controller
      */
     public function update(Request $request, Thread $thread)
     {
+        if(auth()->user()->id!== $thread->user_id){
+            abort(401,"Unauthorized access!");
+        }
         $this->validate($request,[
             'subject'=>'required|min:10',
             'type'=>'required',
             'thread'=>'required|min:20'
         ]);
+
         $thread->update($request->all());
         return redirect()->route('thread.show',$thread->id)->withMessage('Thread Updated!');
     }
@@ -101,6 +105,9 @@ class ThreadController extends Controller
      */
     public function destroy(Thread $thread)
     {
+        if(auth()->user()->id!== $thread->user_id){
+            abort(401,"Unauthorized access!");
+        }
         $thread->delete();
         return redirect()->route('thread.index')->withMessage('Threade is Deleted!');
     }
